middleware/kubernetes: put reverse function in own file
Move reverse function and (some) test code to reverse*.go
